[#47019] Replace LdapGroups cells by ViewComponent
https://community.openproject.org/work_packages/47019
